+++ src/components/HomePage/Section2/Section2.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import car4 from '../../../assets/car/car4.jpg'
+// import car4 from '../../../assets/car/car4.jpg'
 
 const Section2 = () => {
     return (
         <div className="card lg:card-side bg-base-100 shadow-xl rounded-none">
-            <figure><img src={car4} alt="Album" className='' /></figure>
+            {/* <figure><img src={car4} alt="Album" className='' /></figure> */}
             <div className="card-body w-2/3">
                 <h2 className="card-title">New album is released!</h2>
                 <p>Click the button to listen on Spotiwhy app.</p>
