+++ src/components/Layout/Main.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Navbar from '../components/shared/Navbar/Navbar';
+import Navbar from '../shared/Navbar/Navbar';
 
 const Main = () => {
     return (
         <>
-            <Navbar></Navbar>
-            <Outlet></Outlet>
+            <Navbar />
+            <Outlet />
         </>
     );
 };
