+++ src/components/shared/Booking_Service/Booking_Service.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Booking_Service() {
+    return (
+        <div>
+            <h1>Booking_Service section</h1>
+        </div>
+    )
+}
